# swagger-ui/openapi.yaml
...
openapi: 3.0.0
info:
  title: SensorTag IoT API
  description: >-
    API documentada con Swagger usando OpenAPI 3.0.0. con Autenticación Bearer JWT. Desarralla usando NodeJS y el Framework Express de la paquetería de NPM.
    
    
    Documentación extra:
    
    - [OpenAPI Specitification V 3.0.0](https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.3.md#format)
    
    - [NPM](https://www.npmjs.com/)
    
    - [MongoDB](https://www.mongodb.com/)
    
    
    
    
  termsOfService: 'http://swagger.io/terms/'
  contact:
    name: 'Homero Torrijos'
    url:  'https://www.linkedin.com/in/homerotorrijos/'
    email: 'homertor@ucm.es'
  license:
    name: MIT Licence
    url: 'https://opensource.org/licenses/MIT'
  version: "1.0.0"
...
servers:
  - url: https://localhost:3005/
    description: Certificado SSL
  - url: http://localhost:3005/
    description: Sin certificado SSL
...
paths:
...
  /api/v1/refreshTokens:
    get:
      summary: Muestra los refreshTokens 
      description: Lista los refreshTokens con un id_user que se ha logeado correctamente. Esto solo lo pueden ver los administradores
      operationId: getRefreshTokens
      security: 
        - bearerAuth: []
      tags:
        - Users
      responses:
        '200':
          description: Muestra los usuarios.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/refreshToken'
        '404':
          description: No existe el recurso
        '500':
          description: Error al hacer petición
  /api/v1/refreshTokens/{refreshTokenId}:
    delete:
      summary: Elimina un registro con el id del refreshToken
      description: Con el _id generado en la base de datos en MongoDB se elimina un registro. Esto solo lo pueden hacer los administradores
      operationId: refreshTokenId
      security:
        - bearerAuth: []
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/ParamRefreshTokenId'
      responses:
        '200':
          description: El refreshToken ha sido eliminado
        '403':
          description: Sin autorización
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthentication'
        '500':
          description: Error al borrar el refreshToken
  /api/v1/token:
    post:
      summary: Devuelve un token de acceso.
      description: Necesitar tener un pocesión de tu Token de refresco.
      operationId: createToken
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                refreshToken:
                  description: Token de refresco generado por el sistema la primera vez que se logeo.
                  type: string
                  example: 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI1ZjNmNjI2MWY0YzZhMzc5MWM2Yjg0YmMiLCJyb2xlIjoiYWRtaW4iLCJpYXQiOjE1OTc5OTAwMzV9.MkwVp_wBCFZV5A3uBIb_W87an1DdKtyjsNUsEy04-C4'
              required:
                - refreshToken
            encoding:
              color:            # color=red,green,blue
                style: form
                explode: false
          application/json:
            schema:
              $ref: '#/components/schemas/Token'  
      
      #parameters:
      #  - $ref: '#/components/parameters/ParamSensorTag'
      #  - $ref: '#/components/parameters/ParamSensorValor'
      #  - $ref: '#/components/parameters/ParamSensor'
      #  - $ref: '#/components/parameters/ParamSensorDate'
      
      responses:
        '200':
          description: Muestra el sensor registrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseToken'
        '403':
          description: No tienes autorización
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthentication'
        '404':
          description: refreshToken no encontrado
        '500':
          description: Error al hacer petición
...
  /api/v1/users:
    get:
      summary: Muestra los usuarios registrados
      description: Muestra todos los usuarios
      operationId: getUsers
      security: 
        - bearerAuth: []
      tags:
        - Users
      responses:
        '200':
          description: Muestra los usuarios.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          description: No tienes autorización
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthentication'
        '404':
          description: No existe el recurso
        '500':
          description: Error al hacer petición   
  /api/v1/sensortag:
    get:
      summary: Obtiene todos los sensores
      description: Muestra cada sensor con sus lecturas correspondientes
      operationId: getSensors
      #deprecated: true
      security:
        - bearerAuth: []
      
      tags:
        - Sensor
      responses:
        '200':
          description: Muestra los usuarios.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sensor'
        '403':
          description: No tienes autorización
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthentication'
        '404':
          description: No existe el recurso
        '500':
          description: Error al hacer petición
    post:
      summary: Registro de nuevos valores sensados
      description: Estos valores son registrados a través de un cliente REST.
      operationId: createSensor
      security:
        - bearerAuth: []
      tags:
        - Sensor
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                sensorTag:          
                  type: string
                  example: sensortag_cc2650
                valor:    
                  type: string
                  example: 24.5
                sensor:
                  type: string
                  example: pre
                fecha:
                  type: string
                  format: date-time
                  example: 2020-09-04T22:00:00.000Z
              required:
                - sensorTag
                - valor
                - sensor
                - fecha
            encoding:
              color:            # color=red,green,blue
                style: form
                explode: false
          application/json:
            schema:
              $ref: '#/components/schemas/Sensor'  
      
      #parameters:
      #  - $ref: '#/components/parameters/ParamSensorTag'
      #  - $ref: '#/components/parameters/ParamSensorValor'
      #  - $ref: '#/components/parameters/ParamSensor'
      #  - $ref: '#/components/parameters/ParamSensorDate'
      
      responses:
        '200':
          description: Muestra el sensor registrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseSensor'
        '403':
          description: No tienes autorización
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthentication'
        '404':
          description: No existe el recurso
        '500':
          description: Error al hacer petición
  /api/v1/sensortag/{sensortagId}:
    get:
      summary: Obtiene datos de un Sensortag
      description: Obtain detailed information on a single account
      operationId: sensortagId
      security:
        - bearerAuth: []
      tags:
        - Sensor
      parameters:
        - $ref: '#/components/parameters/ParamSensorID'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorID'
        '403':
          description: No tienes autorización
        '404':
          description: No existe el recurso
        '500':
          description: Error al hacer petición
    put:
      summary: Actualiza un registro con el Id del Sensortag
      description: Obtain detailed information on a single account
      security:
        - bearerAuth: []
      tags:
        - Sensor
      parameters:
        - $ref: '#/components/parameters/ParamSensorID'
        - $ref: '#/components/parameters/ParamSensorTag'
        - $ref: '#/components/parameters/ParamSensor'
        - $ref: '#/components/parameters/ParamSensorValor'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorID'
        '403':
          description: No tienes autorización
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthentication'
        '500':
          description: Error al hacer petición
    delete:
      summary: Elimina un registro con el Id del Sensortag
      description: Con el _id generado en la base de datos en MongoDB se elimina un registro.
      security:
        - bearerAuth: []
      tags:
        - Sensor
      parameters:
        - $ref: '#/components/parameters/ParamSensorID'
      responses:
        '200':
          description: El sensorTag ha sido eliminado
        '403':
          description: Sin autorización
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthentication'
        '500':
          description: Error al borrar el sensor
components:
...
  parameters:
...
    ParamSensorTag:
      name: sensorTag
      in: query
      description: id del Sensortag.
      required: true
      schema:
        type: string
        example: sensortag_cc2650
    ParamSensor:
      name: sensor
      in: query
      description: ID del sensor temperatura(tem), humedad(hum), luminosidad(lum), presión(pre).
      required: true
      schema:
        type: string
        example: lum
    ParamSensorValor:
      name: valor
      in: query
      description: Lectura del sensor.
      required: true
      schema:
        type: number
        example: 12.4
    ParamSensorDate:
      name: fecha
      in: query
      description: Fecha en el cuál fue registrado.
      required: true
      schema:
        type: string
        format: DateTimeString
    ParamSensorID:
        name: sensortagId
        in: path
        description: >-
          _id del sensor registrado en la base de datos en MongoDB
        required: true
        schema:
          type: string
          format: ASCIIString
          example: '5eb6176ad71d2e0da86f8d62'
    ParamRefreshTokenId:
        name: refreshTokenId
        in: path
        description: >-
          _id registrado en la base de datos en MongoDB
        required: true
        schema:
          type: string
          format: ASCIIString
          example: '5eb6176ad71d2e0da86f8d62'
  schemas:
...
     ResponseToken:
       properties:
        Mensaje:
          example: 'Este es tu accessToken'
          type: string
        accessToken:
          description: Token único por usuario
          example: 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI1ZjNlMjFjMmVlYjU0MDc3OGMyZTYzOWEiLCJyb2xlIjoiZ3Vlc3QiLCJpYXQiOjE1OTgyODI2NTgsImV4cCI6MTU5ODMwNzg1OH0.scjKA44uTw81OW-iZCEhWGHXSzgyWr7Kg6CfusB80hc'
          type: string
     Token:
      properties:
        refreshToken:
          description: Token único por usuario
          example: 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI1ZjNlMjFjMmVlYjU0MDc3OGMyZTYzOWEiLCJyb2xlIjoiZ3Vlc3QiLCJpYXQiOjE1OTc5ODY4NzR9.qmTvGO7rtrzNveFH4v-Ugs-blzUVLp5HFC-Q69kYQx0'
          type: string
     refreshToken:
      properties:
        _id:
          description: _id asignado por MongoDB
          example: '5f3f583a50ad8a74b4845e1c'
          type: string
        id_user:
          description: id_ del usario
          example: '5f3f6261f4c6a3791c6b84bc'
          type: string
        refreshToken:
          description: Token de refresco para crear nuevos tokens de acceso (accessTokens)
          type: string
          example: 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI1ZjNmNjI2MWY0YzZhMzc5MWM2Yjg0YmMiLCJyb2xlIjoiYWRtaW4iLCJpYXQiOjE1OTc5OTAwMzV9.MkwVp_wBCFZV5A3uBIb_W87an1DdKtyjsNUsEy04-C4'
     ResponseSensor:
      type: object
      properties:
        _id:
          description: "_id asignado por MongoDB"
          example: '5f3d45c43a1bf445846cbf76'
          type: string
        sensor:
          example: 'hum'
          enum: 
            - hum
            - pre
            - lum
            - tem]
          type: string
        fecha:
          example: '2020-03-19T23:00:00.000Z'
          type: string
          format: date-time
        valor:
          type: number
          example: 14.5
...
     User:
      required:
        - name
        - email
        - password
        - role
      properties:
        _id:
          type: string
          description: >-
            Nombre del usuario.
          example: '5eb70b7bf929a6556c83a260'
        name:
          type: string
          description: >-
            Nombre del usuario.
          example: 'Juan'
        password:
          type: string
          description: >-
            Contraseña usada para autenticarse.
          example: '87s?sd!'
        email:
          type: string
          description: >-
            Formato de correo eletrónico muy bien definido, este se encuentra definido como addr_spec en el RFC_5322.
          example: 'mail@mail.com'
        singupDate:
          type: string
          pattern: '^\d{2}-\d{2}-\d{4}$'
          description: >-
            Fecha del usuario en la cuál es creado.
          format: DateTimeString
          example: '12-10-2020'
        role:
          type: string
          enum:
            - admin
            - member
          example: 'admin'
     #Apartado Sensor
     Sensor:
      required:
        - sensorTag
        - sensor
        - valor
        - fecha
      properties:
        #_id:
        #  type: string
        #  description: >-
        #    Es es automaticamente asignado por la base de datos (MongoDB).
        #  example: '5eb6176ad71d2e0da86f8d62'
        sensor:
          type: string
          description: >-
            ID del sensor temperatura(tem), humedad(hum), luminosidad(lum), presión(pre).
          enum:
            - tem
            - hum
            - lum
            - pre
          example: 'tem'
        sensorTag:
          type: string
          description: >-
            ID del sensor temperatura(tem), humedad(hum), luminosidad(lum), presión(pre).
          example: 'sen_cc2650_home'
        valor:
          type: number
          description: >-
            Fecha del usuario en la cuál fue su ultimo acceso.
          example: 14.2
        fecha:
          type: string
          pattern: '^\d{2}-\d{2}-\d{4}$'
          description: >-
            Fecha del usuario en la cuál fue su último acceso. 'MM-DD-YYYY'
          format: DateTimeString
          example: '2020-09-04T22:00:00.000Z'
     #Apartado Autenticación
...
     NotAuthentication:
      type: object
      properties:
        Mensaje:
          type: string
          example: 'No tienes autorización'
     SensorID:
      type: object
      required:
        - _id
      properties:
        _id:
          type: string
          description: >-
            Es es automaticamente asignado por la base de datos (MongoDB).
          example: '5eb6176ad71d2e0da86f8d62'
  securitySchemes:
    bearerAuth:           
      type: http
      scheme: bearer
      bearerFormat: JWT